requirements.txt
keyboard==0.13.5
ndspy==4.1.0
numpy==1.26.1
opencv-python==4.8.1.78
py-desmume==0.0.6
pycairo==1.26.0
pygame==2.5.2
pywin32==306
pillow~=10.3.0